+++ circuit-codegen-annotations/src/nativeMain/kotlin/com/slack/circuit/codegen/annotations/CircuitInject.native.kt
@@ -1,18 +1,18 @@
-// Copyright (C) 2022 Slack Technologies, LLC
+// Copyright (C) 2024 Slack Technologies, LLC
 // SPDX-License-Identifier: Apache-2.0
 package com.slack.circuit.codegen.annotations
 
 import com.slack.circuit.runtime.screen.Screen
-import dagger.hilt.GeneratesRootInput
 import kotlin.reflect.KClass
+import software.amazon.lastmile.kotlin.inject.anvil.extend.ContributingAnnotation
 
 /**
- * JVM-specific [CircuitInject], applying the Hilt [GeneratesRootInput] annotation.
+ * Native-specific [CircuitInject].
  *
  * For more general information about this annotation, see
  * [com.slack.circuit.codegen.annotations.CircuitInject]
  */
-@GeneratesRootInput
+@ContributingAnnotation
 @Target(AnnotationTarget.CLASS, AnnotationTarget.FUNCTION)
 public actual annotation class CircuitInject(
   actual val screen: KClass<out Screen>,