+++ CHANGELOG.md
@@ -4,10 +4,25 @@ Changelog
 **Unreleased**
 --------------
 
+0.24.0
+------
+
+_2024-09-16_
+
+- **New**: Add code gen support for [kotlin-inject](https://github.com/evant/kotlin-inject) + [kotlin-inject-anvil](https://github.com/amzn/kotlin-inject-anvil). See the [code gen docs](https://slackhq.github.io/circuit/code-gen/) for usage instructions. We've also added a sample project.
+- **New**: `presenterTestOf()` and `Presenter.test()` functions now return a new `CircuitReceiveTurbine` interface. This interface slightly but importantly modifies the behavior of `awaitItem()` by making it only emit _changed_ items rather than every item. If you do want to assert the equivalent state is emitted across recompositions, you can use `awaitUnchanged()`.
+- **New**: Promote `LocalBackStack` to public API.
+- **Behavior change**: Due to the above-documented change to `awaitItem()`, you may need to update tests that previously assumed duplicate emissions.
 - Update to Kotlin `2.0.20`.
 - **Change**: Switch to stdlib's implementation of `Uuid`. This release now requires Kotlin `2.0.20` or later.
-- **Behavior change**: `presenterTestOf` and `Presenter.test` have a new optional `moleculeFlowTransformer` parameter that allows for advanced filtering of the `Flow` returned out of the underlying `moleculeFlow`. By default, this now runs a `distinctUntilChanged` operator.
-- Build against KSP `2.0.20-1.0.24`.
+- Add sample test to demonstrate `rememberAnsweringNavigator` result handling.
+- Update to kotlinx.coroutines `1.9.0`.
+- Update to compose-bom `2024.09.01`.
+- Update to kotlinx.collections.immutable `0.3.8`.
+- Update to androidx.activity `1.9.2`.
+- Update to androidx.lifecycle `2.8.5`.
+- Update to kotlin `2.0.20`.
+- Build against KSP `2.0.20-1.0.25`.
 
 0.23.1
 ------